+++ serialisation/src/test/java/org/ehrbase/serialisation/dbencoding/DBEncodeTest.java
@@ -798,4 +798,31 @@ public class DBEncodeTest {
         assertEquals(4, ((Map<String, Object>) feederMap.get("other_details[openEHR-EHR-ITEM_TREE.generic.v1]")).size());
     }
 
+    @Test
+    public void compositionEncodingEmptyProtocol() throws Exception {
+        String value = IOUtils.toString(CompositionTestDataCanonicalJson.GECCO_PERSONENDATEN.getStream(), UTF_8);
+        CanonicalJson cut = new CanonicalJson();
+        Composition composition = cut.unmarshal(value, Composition.class);
+
+        assertNotNull(composition);
+
+        CompositionSerializer compositionSerializerRawJson = new CompositionSerializer();
+
+        String db_encoded = compositionSerializerRawJson.dbEncode(composition);
+        assertNotNull(db_encoded);
+
+        String converted = new LightRawJsonEncoder(db_encoded).encodeCompositionAsString();
+
+        assertNotNull(converted);
+
+        //see if this can be interpreted by Archie
+        Composition object = new CanonicalJson().unmarshal(converted, Composition.class);
+
+        assertNotNull(object);
+
+        String interpreted = new CanonicalXML().marshal(object);
+
+        assertNotNull(interpreted);
+    }
+
 }