+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 - Generated Dtos: Enable usage of Dots with inheritance  (see https://github.com/ehrbase/openEHR_SDK/pull/244)
 - Fix validation of normal statuses (see https://github.com/ehrbase/openEHR_SDK/pull/249)
 - Fix terminology validation issue (see https://github.com/ehrbase/openEHR_SDK/pull/263)
+- Fix NullPointerException when `protocol.items` is `null` or empty (see https://github.com/ehrbase/openEHR_SDK/pull/262)
 
 ## 1.5.0